Delete tables on re-run that have been previously created for experiment
